MVC architecture implemented
